refactor(instance): split by concern (Phase 7)
Implements Phase 7 of the Rust structure refactor roadmap. Splits
`src/instance.rs` (982 LOC, security-sensitive auth-forwarding and
container-state preparation) into three concern-focused files plus
a thin orchestration `mod.rs`.

The plan's rationale: this file contains auth-forward logic
(`ignore`/`copy`/`sync` modes), host-credential reading, macOS
Keychain fallback, symlink rejection, and private-file-permission
repair. Smaller files reduce accidental breakage risk when future
auth changes land.

Split:

- `mod.rs` (124 LOC, was 982) — `AuthProvisionOutcome` enum,
  `AgentState` struct, orchestration (`prepare` / `tear_down`),
  plugin-json serialization (left inline, minimal churn), and
  re-exports of the naming helpers.
- `naming.rs` (63 LOC) — `runtime_slug`, `primary_container_name`,
  `next_container_name`, `class_family_matches`. All four stay `pub`
  (external callers in `src/app/`, `src/runtime/`).
- `auth.rs` (737 LOC) — `impl AgentState { fn provision_claude_auth }`
  (split `impl` block; function body verbatim) plus five private
  helpers: `copy_host_claude_json`, `read_host_credentials`,
  `reject_symlink`, `write_private_file`, `repair_permissions`.
- `plugins.rs` (111 LOC) — `PluginState<'a>` struct (used by
  `mod.rs::AgentState::prepare` to write `plugins.json`).

Verbatim moves throughout. No logic change. No permission-bit edits.
No signature changes. Every symlink guard, every private-file check,
every Keychain fallback path unchanged.

Placement choice for `provision_claude_auth`: Option B (split `impl
AgentState` block in `auth.rs`). Kept as a method to preserve the
exact `Self::provision_claude_auth(...)` call site in
`AgentState::prepare` — no reshape into a free function.

Plugin-json serialization: left inline in `AgentState::prepare`
(choice (a), minimal churn). `plugins.rs` holds only the
`PluginState<'a>` struct; `prepare` imports it via
`use plugins::PluginState`.

Test Preservation Policy: all 18 tests preserved, co-located:

- `mod.rs` — 1 orchestration test (`prepares_persisted_claude_state`).
- `naming.rs` — 2 naming tests (`picks_next_clone_name`,
  `distinguishes_namespaced_and_flat_class_container_names`).
- `auth.rs` — 13 auth-mode and symlink tests (every `ignore_mode_*`,
  `copy_mode_*`, `sync_mode_*`, `rejects_symlink_*`,
  `switching_*_revokes_forwarded_credentials`,
  `auth_file_has_restricted_permissions`,
  `sync_repairs_permissions_*`).
- `plugins.rs` — 2 plugin serialization tests.

No test deleted. No test weakened. No assertion dropped.

Visibility changes (two bumps, both scoped `pub(super)`):
- `AgentState::provision_claude_auth` — was private, now `pub(super)`.
- `PluginState<'a>` + its two fields — were private, now `pub(super)`.

No public API widened. All 12 external references (in `src/app/`,
`src/runtime/`) resolve unchanged. Zero call sites outside
`src/instance/` required edits.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 418 tests run: 418 passed, 0 skipped (unchanged)
